alphacamper-api/pyproject.toml
[project]
name = "alphacamper-api"
version = "0.1.0"
description = "Campflare-style campsite availability and alerts API for Alphacamper"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.16.0,<2.0.0",
  "celery>=5.4.0,<6.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "geoalchemy2>=0.16.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "python-dateutil>=2.9.0,<3.0.0",
  "redis>=5.2.0,<6.0.0",
  "sqlalchemy>=2.0.40,<3.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "ruff>=0.11.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

